requirements.txt
absl-py==2.3.1
aioschedule==0.5.2
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
asyncio==4.0.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
blis==0.7.11
boto3==1.40.57
botocore==1.40.57
cachelib==0.13.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cryptography==46.0.3
cssselect==1.3.0
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.14
decorator==5.2.1
deepface==0.0.95
defusedxml==0.7.1
distro==1.9.0
executing==2.2.0
fastapi==0.104.1
fastjsonschema==2.21.1
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.0
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
fonttools==4.58.5
fqdn==1.5.1
fsspec==2025.9.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
gTTS==2.5.4
gunicorn==21.2.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.3.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server-mathjax==0.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_git==0.51.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.11.3
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lxml==6.0.2
lxml_html_clean==0.4.3
lz4==4.4.4
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mcp==1.18.0
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.1.3
ml_dtypes==0.5.3
mpmath==1.3.0
msgspec==0.19.0
mtcnn==1.0.0
murmurhash==1.0.13
namex==0.1.0
narwhals==1.46.0
nbclient==0.10.2
nbconvert==7.16.6
nbdime==4.0.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.9.2
notebook_shim==0.2.4
numpy==1.26.4
openai==1.55.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
playsound==1.3.0
plotly==6.2.0
pluggy==1.6.0
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==4.25.8
psutil==5.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
PyAudio==0.2.14
pycparser==2.22
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.0.0
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
requests-file==3.0.1
retina-face==0.0.17
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.2.0
rpds-py==0.26.0
s3transfer==0.14.0
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.1
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.4.1
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.3
soupsieve==2.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.44
srsly==2.5.1
srt==3.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.27.0
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
termcolor==3.1.0
terminado==0.18.1
tf_keras==2.19.0
thinc==8.2.5
threadpoolctl==3.6.0
tinycss2==1.4.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.22.1
torch==2.7.1+cpu
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vosk==0.3.45
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0

# Added during Derek Exceptional Upgrade
soundfile==0.13.1
librosa==0.11.0
